refactor: move storage fuctions into submodules
